requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pillow>=8.3.1
numpy>=1.21.0
pyyaml>=5.4.1
einops>=0.4.1
timm>=0.6.12
opencv-python>=4.5.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
prometheus-client>=0.11.0
pytest>=6.2.5
pytest-cov>=2.12.1
httpx>=0.18.2
python-dotenv>=0.19.0
scikit-learn>=0.24.2
pandas>=1.3.0
matplotlib>=3.4.3
seaborn>=0.11.2
tqdm>=4.62.2
